Improve pretty-printing to support invalid bitflags
This may occur when values are reinterpreted from raw integers.
Rather than printing a chain of ORs which do not sum to the explicit
value (or worse yet, print nothing for a disjoint bitset), just tack
on to the end of the chain the explicit numerical value of all
unhandled bits.

We can even represent this in the non-compact case as a reinterpret
expression which is copy-pastable to re-form the given value.