Test functionality of automatic conversion in ccall
This adds tests for the `cconvert` method used implicitly during
`ccall`s. When designing the tests, though, it became obvious that
real-world conversion to `Cint` arguments (such as calls to `mmap`
in UnixMmap.jl) weren't covered by the existing method.

Change the behavior to allow for conversion to any integer type which
can be represented by the destination type.