Fix doc string tests in prep for Julia v1.11 release
The change in JuliaLang/julia#54499 fixed the ability to access/copy
doc strings without needing to take an explicit dependency on the
REPL package, but the returned object does not trivially reduce back
to a string, as described in issue JuliaLang/julia#54664.

Instead of waiting for that issue to get resolved at the Julia level
to fix our tests, just add a helper shim to the tests to access the
desired information. The goal is to just ensure that the macro
`@__doc__` expression is correctly used to permit documenting the
generated type definitions, so we don't really care about the specific
form as long as Julia uses it correctly within the docs system.